# api/docs/openapi.yaml
openapi: 3.1.0
info:
  title: Solar Energy Assistant API
  version: 1.0.0
  description: Local single-owner API for owner onboarding, installations, datasets, training, predictions, dashboard, and recommendations.
servers:
  - url: http://localhost:8000
...
paths:
...
  /installations/active:
    get:
      tags: [installations]
      operationId: get_active_installation
      responses:
        '200':
          description: Active installation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InstallationConfigurationResult'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /installations/{installation_id}/configuration:
    get:
      tags: [installations]
      operationId: get_installation_configuration
      parameters:
        - $ref: '#/components/parameters/InstallationId'
      responses:
        '200':
          description: Installation configuration
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InstallationConfigurationResult'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      tags: [installations]
      operationId: save_installation_configuration
      parameters:
        - $ref: '#/components/parameters/InstallationId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InstallationConfigurationUpdate'
      responses:
        '200':
          description: Installation configuration updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InstallationConfigurationResult'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
  /installations/{installation_id}/dashboard:
    get:
      tags: [installations]
      operationId: get_installation_dashboard
      parameters:
        - $ref: '#/components/parameters/InstallationId'
      responses:
        '200':
          description: Dashboard summary
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DashboardResult'
  /installations/{installation_id}/predictions:
    get:
      tags: [installations]
      operationId: predict_installation_solar_yield
      parameters:
        - $ref: '#/components/parameters/InstallationId'
        - $ref: '#/components/parameters/StartDate'
        - $ref: '#/components/parameters/EndDate'
      responses:
        '200':
          description: Prediction result or not-ready state
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/PredictionResult'
                  - $ref: '#/components/schemas/PredictionNotReadyResult'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
  /datasets:
    get:
      tags: [datasets]
      operationId: list_datasets
      parameters:
        - in: query
          name: installation_id
          schema:
            type: string
        - in: query
          name: limit
          schema:
            type: integer
            default: 100
            minimum: 1
            maximum: 500
        - in: query
          name: offset
          schema:
            type: integer
            default: 0
            minimum: 0
      responses:
        '200':
          description: Dataset list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DatasetResult'
    post:
      tags: [datasets]
      operationId: create_dataset
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file, installation_id]
              properties:
                file:
                  type: array
                  items:
                    type: string
                    format: binary
                installation_id:
                  type: string
                mapping:
                  oneOf:
                    - type: string
                    - type: object
                date_column:
                  type: string
                time_column:
                  type: string
                measurement_column:
                  type: string
                unit:
                  type: string
      responses:
        '200':
          description: Dataset uploaded and converted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DataUploadResult'
        '400':
          $ref: '#/components/responses/BadRequest'
  /datasets/inspect:
    post:
      tags: [datasets]
      operationId: inspect_dataset
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
      responses:
        '200':
          description: CSV columns
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CsvColumnsResult'
        '400':
          $ref: '#/components/responses/BadRequest'
  /datasets/{dataset_id}:
    get:
      tags: [datasets]
      operationId: get_dataset
      parameters:
        - in: path
          name: dataset_id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Dataset
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DatasetResult'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  parameters:
    InstallationId:
      in: path
      name: installation_id
      required: true
      schema:
        type: string
...
    StartDate:
      in: query
      name: start_date
      required: true
      schema:
        type: string
        format: date
    EndDate:
      in: query
      name: end_date
      required: true
      schema:
        type: string
        format: date
  responses:
    BadRequest:
      description: Validation or request error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
    ErrorResponse:
      type: object
      properties:
        detail:
          type: string
      required: [detail]
...
    InstallationConfigurationUpdate:
      type: object
      properties:
        name:
          type: string
          nullable: true
        country:
          type: string
          nullable: true
        city:
          type: string
          nullable: true
        panel_count:
          type: integer
          nullable: true
        capacity_kwp:
          type: number
          format: double
          nullable: true
        installation_kwp:
          type: number
          format: double
          nullable: true
        consuming_activities:
          type: array
          items:
            $ref: '#/components/schemas/ConsumingActivity'
        appliances:
          $ref: '#/components/schemas/ApplianceConsumption'
    InstallationConfigurationResult:
      type: object
      required:
        [id, owner_id, installation_id, available_consuming_activities, consuming_activities, appliances]
      properties:
        id:
          type: string
        owner_id:
          type: string
        installation_id:
          type: string
        is_active:
          type: boolean
          nullable: true
        name:
          type: string
          nullable: true
        country:
          type: string
          nullable: true
        city:
          type: string
          nullable: true
        latitude:
          type: number
          format: double
          nullable: true
        longitude:
          type: number
          format: double
          nullable: true
        panel_count:
          type: integer
          nullable: true
        capacity_kwp:
          type: number
          format: double
          nullable: true
        available_consuming_activities:
          type: array
          items:
            $ref: '#/components/schemas/ConsumingActivity'
        consuming_activities:
          type: array
          items:
            $ref: '#/components/schemas/ConsumingActivity'
        appliances:
          $ref: '#/components/schemas/ApplianceConsumption'
    ApplianceConsumption:
      type: object
      properties:
        car:
          type: number
          format: double
        washing_machine:
          type: number
          format: double
        dishwasher:
          type: number
          format: double
        dryer:
          type: number
          format: double
        boiler:
          type: number
          format: double
    ConsumingActivity:
      type: object
      required: [name, consumption_kwh]
      properties:
        id:
          type: string
          nullable: true
        name:
          type: string
        consumption_kwh:
          type: number
          format: double
        category:
          type: string
          nullable: true
        is_custom:
          type: boolean
    DashboardResult:
      type: object
      required: [installation_id, current_weather, tomorrow_weather]
      properties:
        installation_id:
          type: string
        current_output_kw:
          type: number
          format: double
          nullable: true
        todays_yield_kwh:
          type: number
          format: double
          nullable: true
        todays_peak_time_window:
          type: string
          nullable: true
        current_weather:
          $ref: '#/components/schemas/WeatherSummaryResult'
        tomorrow_weather:
          $ref: '#/components/schemas/TomorrowWeatherResult'
    WeatherSummaryResult:
      type: object
      properties:
        temperature_c:
          type: number
          format: double
          nullable: true
        status:
          type: string
          nullable: true
    TomorrowWeatherResult:
      type: object
      properties:
        min_temperature_c:
          type: number
          format: double
          nullable: true
        max_temperature_c:
          type: number
          format: double
          nullable: true
        status:
          type: string
          nullable: true
    DatasetResult:
      type: object
      required: [id, installation_id, dataset_hash, path, row_count, granularity, value_column, created_at]
      properties:
        id:
          type: string
        installation_id:
          type: string
        dataset_hash:
          type: string
        path:
          type: string
        row_count:
          type: integer
        granularity:
          type: string
        value_column:
          type: string
        production_unit:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
    CsvColumnsResult:
      type: object
      required: [filename, columns]
      properties:
        filename:
          type: string
        columns:
          type: array
          items:
            type: string
        preview_rows:
          type: array
          items:
            $ref: '#/components/schemas/CsvPreviewRowResult'
    CsvPreviewRowResult:
      type: object
      required: [values]
      properties:
        values:
          type: object
          additionalProperties:
            type: string
            nullable: true
    TimeResolutionResult:
      type: object
      required: [kind]
      properties:
        kind:
          type: string
        interval_minutes:
          type: number
          format: double
          nullable: true
    DataUploadResult:
      type: object
      required:
        [dataset_id, dataset_hash, dataset_already_exists, installation_id, production_unit, rows, columns, value_column, time_resolution, parquet_path]
      properties:
        dataset_id:
          type: string
        dataset_hash:
          type: string
        dataset_already_exists:
          type: boolean
        installation_id:
          type: string
        production_unit:
          type: string
        rows:
          type: integer
        columns:
          type: array
          items:
            type: string
        value_column:
          type: string
        time_resolution:
          $ref: '#/components/schemas/TimeResolutionResult'
        parquet_path:
          type: string
...
    PredictionPointResult:
      type: object
      required: [timestamp, value]
      properties:
        timestamp:
          type: string
          format: date-time
        value:
          type: number
          format: double
    PredictionResult:
      type: object
      required: [unit, total_average, predictions]
      properties:
        unit:
          type: string
        total_average:
          type: number
          format: double
        predictions:
          type: array
          items:
            $ref: '#/components/schemas/PredictionPointResult'
    PredictionNotReadyResult:
      type: object
      required: [status, reason, training_run_id]
      properties:
        status:
          type: string
        reason:
          type: string
        training_run_id:
          type: string
...